Use TIMESTAMP_TZ types when creating the table
Previously we used the `TIMESTAMP` column type. In Snowflake, this is
_probably_ an alias for `TIMESTAMP_NTZ`, but actually it depends on
whether the account owner has change the default of the
`TIMESTAMP_TYPE_MAPPING` session parameter.

We should pick a concrete timestamp type and use it explicitly for all
new tables. This will prevent future confusion where we are not 100%
sure what timestamp got used when creating the table.

This change was supposed to happen in the first release, but it got
overlooked.