Add e2e_latency_millis metric (#69)
This PR adds a new metric `e2e_latency_millis` which measures the
difference between an event's `collector_tstamp` and when it was
written to Snowflake.

The metric is emitted to statsd only on minutes in which some
events are written to Snowflake.

The metric measures the worst-case latency for a batch, i.e. latency
for the earliest seen `collector_tstamp`.