Use offset tokens to ensure at-least-once semantics
Until now, we have (effectively) used response codes from Snowflake to
tell us that the events are safely in Snowflake. This was done via the
snowflake java sdk; specifically, if `.flush()` returns with no error
then we assumed the events are in snowflake.

But we hit edge cases where snowflake would response with a successful
response code, but the events were not fully ingested into the table.
This edge case can happen around the time of schema evolution.

After this commit, we use offset tokens to check the ingestion of a
batch of events. We do not assume the batch is ingested to the table
until `channel.getLatestCommittedOffsetToken()` returns the expected
offset.